backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
faster-whisper>=1.0.0
ffmpeg-python==0.2.0
pydantic==2.5.3
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
deep-translator==1.11.4
onnxruntime>=1.16.0
nltk==3.8.1
torch>=2.0.0
huggingface_hub
# Speaker Diarization (optional - requires HF_TOKEN env var)
pyannote.audio>=3.1.0
speechbrain>=0.5.16
